#434 reformat tests
