+++ tests/integration/test_models/test_lead_acid/test_newman_tiedemann.py
@@ -29,6 +29,35 @@ class TestLeadAcidNewmanTiedemann(unittest.TestCase):
         np.testing.assert_array_almost_equal(original, simp_and_known)
 
 
+class TestLeadAcidNewmanTiedemannCapacitance(unittest.TestCase):
+    def test_basic_processing(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.NewmanTiedemann(options)
+        modeltest = tests.StandardModelTest(model)
+        modeltest.test_all()
+
+    @unittest.skipIf(scikits_odes_spec is None, "scikits.odes not installed")
+    def test_basic_processing_no_capacitance(self):
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.NewmanTiedemann(options)
+        modeltest = tests.StandardModelTest(model)
+
+        modeltest.test_all()
+
+    def test_optimisations(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.NewmanTiedemann(options)
+        optimtest = tests.OptimisationsTest(model)
+
+        original = optimtest.evaluate_model()
+        simplified = optimtest.evaluate_model(simplify=True)
+        using_known_evals = optimtest.evaluate_model(use_known_evals=True)
+        simp_and_known = optimtest.evaluate_model(simplify=True, use_known_evals=True)
+        np.testing.assert_array_almost_equal(original, simplified, decimal=5)
+        np.testing.assert_array_almost_equal(original, using_known_evals)
+        np.testing.assert_array_almost_equal(original, simp_and_known, decimal=5)
+
+
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys
